requirements.txt
aiohttp
aioconsole
coloredlogs
uvicorn
fastapi
python-dotenv
